Cython (#7)
* Wrap all functions with Cython.

* Fixed interface to DAFoam.

* Added a new test for basic IO functions.

+++ python/reg_tests/run_reg_tests_cs.py
@@ -28,6 +28,10 @@ diff_cmd = args.diff_cmd
 nodiff = args.nodiff
 mpiexec= args.mpiexec
 
+if os.getenv("WM_PROJECT") is None:
+    print("OpenFOAM environment not found, forgot to source the OpenFOAM bashrc?")
+    exit(1)
+
 if mode == 'train':
     try:
         os.remove('%s_reg.ref'%(module_name))
